chore: rename store from src/lib to src/db
