+++ packages/astro/e2e/actions-blog.test.ts
@@ -15,7 +15,7 @@ test.afterAll(async () => {
 
 test.afterEach(async ({ astro }) => {
 	// Force the in-memory store to re-seed between tests by touching its module
-	await astro.editFile('./src/lib/store.ts', (original) => original, false);
+	await astro.editFile('./src/db/store.ts', (original) => original, false);
 });
 
 test.describe('Astro Actions - Blog', () => {
