+++ packages/astro/e2e/fixtures/actions-blog/src/pages/blog/[...slug].astro
@@ -2,7 +2,7 @@
 import { type CollectionEntry, getEntry, render } from 'astro:content';
 import BlogPost from '../../layouts/BlogPost.astro';
 import { Logout } from '../../components/Logout';
-import { getComments, getLikes } from '../../lib/store';
+import { getComments, getLikes } from '../../db/store';
 import { Like } from '../../components/Like';
 import { PostComment } from '../../components/PostComment';
 import { actions } from 'astro:actions';
